testing-gathering/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alpaca-py==0.21.1
alpha-vantage==2.3.1
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
finnhub-python==2.4.19
fmpsdk==20240330.0
frozendict==2.4.4
frozenlist==1.4.1
html5lib==1.1
idna==3.7
iniconfig==2.0.0
lxml==5.2.2
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
numpy==1.26.4
packaging==24.0
pandas==2.2.2
peewee==3.17.5
platformdirs==4.2.2
pluggy==1.5.0
pyarrow==16.1.0
pydantic==2.7.1
pydantic_core==2.18.2
pytest==7.4.4
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sseclient-py==1.8.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
yfinance==0.2.40
